Formatter and more warning fixes

+++ src/diprint.c
@@ -1,8 +1,8 @@
 #include "diprint.h"
 #include "common.h"
+#include "gameVi.h"
 #include "libc/string.h"
 #include "stdarg.h"
-#include "gameVi.h"
 
 /************ .data ************/
 
@@ -971,7 +971,7 @@ void diPrintfAll(Gfx **dList) {
     gDebugScreenHeight = height;
     gDPSetScissor((*dList)++, 0, 0, 0, gDebugScreenWidth, gDebugScreenHeight);
     debug_text_bounds();
-    gSPDisplayList((*dList)++, dDebugFontSettings);
+    gSPDisplayList((*dList)++, (u32) dDebugFontSettings);
     buffer = (char *) gDebugPrintBufferStart;
     debug_text_origin();
     gDebugFontTexture = -1;
